logging config moved to main widget
